feat: extract move-asset-audio into Action with colocated DTOs
