+++ app/application/api_controller/actions/move_asset_audio/move_asset_audio_request.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class MoveAssetAudioRequest(BaseModel):
+    user_id: str
+    source_asset_id: str
+    target_asset_id: str
