+++ app/application/api_controller/actions/move_asset_audio/move_asset_audio_response.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+
+
+class MovedFile(BaseModel):
+    source_path: str
+    file_path: str
+    file_name: str
+
+
+class MoveAssetAudioResponse(BaseModel):
+    moved: list[MovedFile]
+    count: int
